network: Replace custom OpenFlow with localport OVN configuration
This removes the custom OpenFlow code used to transport packets
out of the OVN private network destined for the public internet
(via IPTables) and the minion (for DNS). The same functionality
is instead implemented purely with OVN config using the new
'localport' feature. This should be significantly easier to
understand and maintain than the custom OpenFlow.